Evalua_V1/requirements.txt
fastapi==0.115.5
uvicorn==0.30.3
python-dotenv==1.0.1
jinja2==3.1.4
sqlalchemy==2.0.25
pydantic==2.7.4
email-validator>=2.0.0
pydantic-settings==2.3.0
python-multipart==0.0.9
aiofiles==23.2.1
psycopg[binary]==3.2.1
openpyxl==3.1.5
bcrypt>=4.0.1,<5
itsdangerous==2.2.0
pytest==8.3.4
httpx==0.28.1
